feat(upload): restructure routing and refactor shared components for upload and reupload pages
- Move route from `page/upload` to `page/(upload)` for better grouping
- Relocate existing upload page and add new reupload page
- Update `DevNavigation` paths to match new routing structure
- Add `UploadTemplate` component for shared layout between upload and reupload pages
- Implement `disabled` and conditional content display logic
- Fix validation logic and adjust final confirmation dialog content
- Refine `PopoverCalendar` component for improved usability and integration
