+++ packages/tv-ad-admin-next/app/(upload)/upload/page.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+import UploadTemplate from '@/components/upload/upload-template'
+
+export default function UploadPage() {
+  return (
+    <UploadTemplate
+      pageTitle="上傳廣告素材"
+      mode="upload"
+      onSubmit={(data) => console.log('送出上傳資料', data)}
+    />
+  )
+}
